[ANDROSDK-1988] move api logic to network layer
